communication: Add guidance on times and timezones
Most mistakes I've made in the past around communications and scheduling
are miscommunications around times and their notation, or timezones
with offsets.

For timezones, most servers are set to UTC for convenience. While this
is not the reason to recommend it, it was the inspiration to suggest it
for the company too. When one goes into summer time, their offset
changes, and only theirs. Putting the responsibility on each and
everyone to remember the offset tot UTC is more convenient for the
entire company.

Around times themself: a 24 hour clock unambiguously communicates when a
meeting is to take place. Using AM/PM is fine though, but has to be
explicitly communicated.