company: Hire for value fit over culture fit
Culture is taught to new hires, and created by a whole team. It's not
static, and thus it's hard to hire for. Values are much more stable and
thus easier to hire for. Trust that if the values are shared, the
culture will follow, and thus hire for value fit.