weekly meeting: Propose new structure
At this company we've got two weekly meetings, this change documents the
structure as is with a proposed new structure for the Tuesday meeting.

The new additions are two fold:
1. Reviewing on the weekend from a personal standpoint
Hoping this will aid in team bonding
2. Open Mic
When ideas are unfinished in ones head, issues aren't created yet.
Having a place to discuss such ideas allows everyone to chip in
improvements to move these ideas along and into issues.