strategy: Explain key metrics to track
While proper tooling is currently missing for FlowForge team members to
track this, the CEO reports on this metrics monthly through the investor
email.

By recording this information in the handbook it can be linked from the
email, and changes to metrics or their definitions can more easily be
made.