communication: Slack is emphemeral
On items discussed on Slack it's usually hard to search and find the
history of the topic. The place that should contain information around
the problem, possible solutions and their tradeoffs, and status reports
should always be the GitHub issue or Google Doc. As such if a decision
is made in Slack it should be recorded to persist this knowledge outside
of Slack.