community: Add a Node-RED consultants page
This is a bare-bones iteration of a list of Node-RED consultants. While
this list is by no means exhaustive or even close to being, I wanted to
record these changes to start collaboration on this list.

The goal is to list known Node-RED consultants for the community to go
to. FlowForge doesn't offer professional services, but we can help
others find them.

This change is missing a navigation route, and about 15 more consultants
before it's ready for publishing. As well as a design overhaul.