kpi: Replace Managed Instances with Platforms
The number of instances is dominated by FlowForge.com while the install
base should also grow in it's breath. By changing the KPI to managed
instances first our north star should change to be more focussed on
breath of adoption.

To be more focussed on this goal the second KPI is dropped. Further, a
bonus is awarded per release the company meets their goals.