misc: Add copy code button
There's often a case on the website where a blog post adds a code block
which users need copy. This change adds a button in the code block to
allow the code block contents to be copied into their clipboard.

This change does add two JS libraries loaded based on a URL, which is
something we might not want to do.

Further; there's a choice of styling. I've chosen the red as it has more
contrast on both the white and dark code block background.