blog: Explain the exec nodes
Recently I got the question why the exec nodes were disabled on
FlowForge Cloud, but there's cases you'd not disable them for each
stack. While this blog post doesn't explain that, it explains the basic
knowledge I assumed others to have for my explaination.

This blog post builds a basis on which to build in follow ups, for
example: why is `stdin` missing? How doe FileDescriptors work, and how
does that influence the exec node? And in what cases do we need the exec
nodes?