status page: add 404 to the root.
Accourding to GH documentation[1], when a 404.html is added to the root
of the repository, it's sending it as a 404 with the same status code.
That would fix a SEO issue for us.

This change is a bit hacky mostly to validate the fix, and later get it
done with a build system.

[1]: https://docs.github.com/en/enterprise-server@3.3/pages/getting-started-with-github-pages/creating-a-custom-404-page-for-your-github-pages-site