docs: Add custom title for page
Before this change all of my docs tabs had the same title when I hovered
on them, "Docs". This wasn't too helpful to select the right one. This
change add the file slug as a prefix. For example, `envvar.md` will now
generate the title as "envvar • FlowForge".

This change also helps with SEO, for indexers <3 title elements.