misc: Fix broken anchor in link
When updating the KPI, see also: cf0d7e4cfc2f571614d723da8fe6574e20c1ec90,
a anchor changed name. This anchor was linked elsewhere, so that link
ended up going to the right page, just not the right section of the
page.

This change updates the link to go to the right anchor.