community: Introduce comments for blog posts
Blog posts are currently a one-way communication channel. With this
change we include the Discourse discussion that's being generated for each
post. Allow a 2 way communication channel.

Currently it's using feature flags as it's not testable at this time.
It needs to be deployed on the website for the embedding to work.