docs: Core nodes can include a use case text
Some docs from upstream make sense to be short in their use case, but
for reference documenation, it makes more sense to have more explicit
details.

This change adds an example for the split node, as I know the node well
to do so, and as-is this change adds an example for more people to be
able to chip in.