rebrand: Internal links should not include domain
While it works to have a full domain name and path in the link, if that
link is internal to our site, it's not required. This change is a
find-and-replace action on the domain to just strip it out.

That makes all links go to FlowFuse instead of our old domain.