testing: Do not allow for capitals in URLs
When the sitemap has URLs with capitals and these URLs are followed,
they get redirected by the webserver. To prevent this we should make
sure that no URL has a capital letter to begin with.

For SEO benefits, as well as for consistency, this change adds a quick
test that validates that for all pages in the Sitemap the URLs do not
contain any capital letters.