docs: Fix core node docs missing Join
There were a couple of problems with the core nodes before this change. First and foremost, some nodes were missing, fore example the join one.

The missing Join node was due the it being in same file as the `split` one. Thus selecting the right file didn't work, and so the node basically didn't exist.

Some cascading changes include a better spelling of the node names, for better human readibility.