fix: Core nodes missing use-case
When this change is deployed the use cases for each core nodes
will be deployed both on filesystems with capital letter sensitivity for
filenames and the ones without.

This change is required as my laptop doesn't care about capital letters and the node.name includes them. As such the use case files were found and rendered. GitHub actions runs a filesystem that's not capital letter 'friendly' and thus didn't include the docs.