Node-RED: Allow flows to be rendered as diagrams
Utilizes the upstream package remove the need for
images + plain JSON.

For this change I've only shown the split node as an
example page. The rest of the nodes will be added
later, though there's some upstream changes that
would be better to add before we roll these changes
out wholesale.