Update for Quote reference on PO
Customer POs are often missing three important pieces of information we need for our subscriptions. 
1. Official start date for the subscription
2. Packaging/feature details within the FlowFuse platform
3. Prices for purchasing additional capacity within the term of the subscription. 

This information is always included in the Company Quote we send to the customer. By requiring the customer to add the quote reference number to their PO we ensure the information above is included in the sale.